refactor: 모달 커스텀 훅 생성
